+++ ReleaseNotes.md
@@ -2,6 +2,12 @@ Q. Never mind the fluff -- how do I get started?
 
 A. Start with the README : https://github.com/SteveGilham/altcode.test/blob/master/README.md
 
+# 1.0.6
+* [BREAKING] `Match` -> `AssertionMatch`
+* `altcode.test.expecto` -- add `Expecto.Flip` support
+* `altcode.test.nunit` -- add `DirectoryAssert`, `FileAssert` and `StringAssert` support
+* `altcode.test.xunit` -- add `AltAssert.Same`
+
 # 1.0.4
 * [NEW PACKAGE] `altcode.test.expecto` -- named argument helper for Expecto >= 8.12.0
 * [NEW PACKAGE] `altcode.test.nunit` -- named argument helper for NUnit >= 3.12.0